BigInteger everything

+++ src/main/java/com/caseykulm/Main.java
@@ -1,7 +1,8 @@
 package com.caseykulm;
 
 import com.caseykulm.primality.PrimalityStrategy;
-import com.caseykulm.primality.SimplePrimality;
+import com.caseykulm.primality.deterministic.SimplePrimality;
+import java.math.BigInteger;
 
 import static com.caseykulm.Utils.basicPrintPrimesToLimit;
 import static com.caseykulm.Utils.logTime;
@@ -9,7 +10,7 @@ import static com.caseykulm.Utils.logTime;
 public class Main {
 
   public static void main(String[] args) {
-    int limit = 100000000;
+    BigInteger limit = BigInteger.valueOf(100000000);
     PrimalityStrategy strategy = new SimplePrimality();
     long timeStart = System.currentTimeMillis();
     basicPrintPrimesToLimit(strategy, limit);
